refactors gameover to make RPCs correctly and work for timer
