+++ Assets/Resources/Enemy.prefab
@@ -519,7 +519,7 @@ Transform:
   m_GameObject: {fileID: 6444308001262241288}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -16.093351, y: -1, z: -43.50849}
+  m_LocalPosition: {x: -19.130112, y: -1, z: -1.7955236}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
