+++ Assets/Resources/Equipment/Sword2.prefab
@@ -30,7 +30,7 @@ Transform:
   m_GameObject: {fileID: 1166987625756554385}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -20, y: -0.90000004, z: 20.6}
+  m_LocalPosition: {x: 40, y: -0.90000004, z: 80.6}
   m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
   m_ConstrainProportionsScale: 1
   m_Children:
