+++ Assets/Resources/Pickups/ManaOrb.prefab
@@ -34,7 +34,7 @@ Transform:
   m_GameObject: {fileID: 1950060490699823590}
   serializedVersion: 2
   m_LocalRotation: {x: -0.7071068, y: 0, z: 0, w: 0.7071068}
-  m_LocalPosition: {x: -19.282629, y: -1.6, z: 19.314022}
+  m_LocalPosition: {x: -39.9603, y: -1.6, z: 59.174255}
   m_LocalScale: {x: 0.25754, y: 0.25754, z: 0.25754}
   m_ConstrainProportionsScale: 0
   m_Children: []
