+++ Assets/Resources/Projectiles/FireProjectile.prefab
@@ -34,7 +34,7 @@ Transform:
   m_GameObject: {fileID: 585884064168877562}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -20.090261, y: -0.4671189, z: -29.338009}
+  m_LocalPosition: {x: -20.985512, y: -0.40952787, z: 13.9534855}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
