+++ Assets/Scripts/Objectives/GameOverSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ca5b5918e00e71f43bf64d4b0f490f0a
+guid: b790275f7eccf4cff99361213942dbde
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
