+++ Assets/Scripts/Player/PlayerHealth.cs
@@ -31,7 +31,7 @@ public class PlayerHealth : NetworkBehaviour
 
         if(currentHealth.Value == 0)
         {
-            UIManager.Instance.gameOver.GameOverServerRpc(false);
+            GameOverSystem.Instance.GameOverServerRpc(false);
         }
 
     }
